+++ src/main/java/com/foru/freebe/reservation/service/CustomerReservationService.java
@@ -10,13 +10,11 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.foru.freebe.common.dto.ApiResponse;
 import com.foru.freebe.errors.errorcode.CommonErrorCode;
-import com.foru.freebe.errors.errorcode.ProductErrorCode;
 import com.foru.freebe.errors.exception.RestApiException;
 import com.foru.freebe.member.entity.Member;
 import com.foru.freebe.member.repository.MemberRepository;
 import com.foru.freebe.product.dto.photographer.ProductComponentDto;
 import com.foru.freebe.product.dto.photographer.ProductOptionDto;
-import com.foru.freebe.product.entity.ActiveStatus;
 import com.foru.freebe.product.entity.Product;
 import com.foru.freebe.product.entity.ProductComponent;
 import com.foru.freebe.product.entity.ProductOption;
@@ -51,6 +49,7 @@ public class CustomerReservationService {
     private final ProductOptionRepository productOptionRepository;
     private final ReferenceImageRepository referenceImageRepository;
     private final S3ImageService s3ImageService;
+    private final ReservationValidator reservationValidator;
 
     public ApiResponse<Long> registerReservationForm(Long id, FormRegisterRequest formRegisterRequest,
         List<MultipartFile> images) throws IOException {
@@ -58,7 +57,7 @@ public class CustomerReservationService {
         Member photographer = findMember(formRegisterRequest.getPhotographerId());
 
         ReservationForm reservationForm = createReservationForm(formRegisterRequest, photographer, customer);
-        validateReservationForm(formRegisterRequest);
+        reservationValidator.validateReservationFormBeforeSave(formRegisterRequest);
 
         List<String> originalImageUrls = s3ImageService.uploadOriginalImages(images, S3ImageType.RESERVATION, id);
         List<String> thumbnailImageUrls = s3ImageService.uploadThumbnailImages(images, S3ImageType.RESERVATION, id,
@@ -104,7 +103,7 @@ public class CustomerReservationService {
         ReservationForm reservationForm = reservationFormRepository.findById(reservationFormId)
             .orElseThrow(() -> new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND));
 
-        validateCustomerAccess(reservationForm, customerId);
+        reservationValidator.validateCustomerAccess(reservationForm, customerId);
 
         ReservationInfoResponse reservationInfoResponse = ReservationInfoResponse.builder()
             .reservationStatus(reservationForm.getReservationStatus())
@@ -156,24 +155,6 @@ public class CustomerReservationService {
         return builder.build();
     }
 
-    private void validateReservationForm(FormRegisterRequest formRegisterRequest) {
-        validateProductTitleExists(formRegisterRequest.getProductTitle());
-        validateProductIsActive(formRegisterRequest.getProductTitle());
-    }
-
-    private void validateProductIsActive(String productTitle) {
-        Product product = productRepository.findByTitle(productTitle);
-        if (product.getActiveStatus() != ActiveStatus.ACTIVE) {
-            throw new RestApiException(ProductErrorCode.PRODUCT_INACTIVE_STATUS);
-        }
-    }
-
-    private void validateProductTitleExists(String productTitle) {
-        if (!productRepository.existsByTitle(productTitle)) {
-            throw new RestApiException(CommonErrorCode.RESOURCE_NOT_FOUND);
-        }
-    }
-
     private List<ProductOptionDto> convertProductOptionDtoList(List<ProductOption> productOptions) {
         List<ProductOptionDto> productOptionDtoList = new ArrayList<>();
         for (ProductOption productOption : productOptions) {
@@ -198,12 +179,5 @@ public class CustomerReservationService {
             productComponentDtoList.add(productComponentDto);
         }
         return productComponentDtoList;
-    }
-
-    private void validateCustomerAccess(ReservationForm reservationForm, Long customerId) {
-        Long reservationCustomerId = reservationForm.getCustomer().getId();
-        if (!reservationCustomerId.equals(customerId)) {
-            throw new RestApiException(CommonErrorCode.ACCESS_DENIED);
-        }
     }
 }
